Add files via upload
A large company pays its salespeople on a commission basis.
The salespeople each receive 150€ per week plus 12% of their gross sales for that week. For example, a salesperson who sells 6000€ worth of chemicals in a week receives 150€ plus 12% of 6000€, or a total of 870€. Develop a C++ program that uses a while statement to input each salesperson’s gross
sales for last week and calculates and displays that salesperson’s earnings. Process one salesperson’s figures at a time.