Add files via upload
Write a C++ program to multiply and then evaluate two polynomials of an arbitrary degree provided by the user. Up to five variables are possible in each polynomial, ex. (10x^10)(15.5y^3)(1.76z^7)(34p^2)(15f^4)+(4y^4)(4p). You are allowed to use power and sqrt functions from C++ standard library along with all input/output libraries (iostream and like so). Other functions/methods are required to be implemented by you. You must follow the object-oriented pattern provided by C++ (using the concept of classes and information hiding at least). Your program should have an adequate user interface (text or graphical), which helps the user to understand and conveniently provides the inputs and understand the output. Employing a graphical user interface is not required, but if implemented, it will get a bonus. The originality of the source code is also of high importance, and plagiarism detection software may apply. After submitting your project, we will have an online or in-person meeting to get more information on the process of developing your program and the level of your programming skills.